Added animation on page transition
