+++ src/App.js
@@ -1,19 +1,19 @@
 import React, { useEffect } from "react";
 import "./App.css";
+import { Switch, Route, useLocation } from "react-router-dom";
+import { useDispatch, useSelector } from "react-redux";
+import { MainNavContainer } from "../src/components/Navigation/index";
+import "typeface-roboto";
+// import "./styles.css";
+import { AnimatePresence, motion } from "framer-motion";
 
-import { Switch, Route } from "react-router-dom";
-// import Navigation from "./components/Navigation";
+import { getUserWithStoredToken } from "./store/user/actions";
+import { selectAppLoading } from "./store/appState/selectors";
+
 import Loading from "./components/Loading";
 import MessageBox from "./components/MessageBox";
 import SignUp from "./pages/SignUp";
 import Login from "./pages/Login";
-// import "./styles.css";
-import { MainNavContainer } from "../src/components/Navigation/index";
-
-import { useDispatch, useSelector } from "react-redux";
-import { selectAppLoading } from "./store/appState/selectors";
-import { getUserWithStoredToken } from "./store/user/actions";
-import 'typeface-roboto';
 import Genres from "./pages/Genres";
 import GenreDetails from "./pages/GenreDetails";
 import MyPage from "./pages/MyPage";
@@ -21,24 +21,30 @@ import MyPage from "./pages/MyPage";
 function App() {
   const dispatch = useDispatch();
   const isLoading = useSelector(selectAppLoading);
+  const location = useLocation();
 
   useEffect(() => {
     dispatch(getUserWithStoredToken());
   }, [dispatch]);
 
   return (
-    <div className="App" style={{backgroundColor: "#f5f5f5", height: "945px"}}>
+    <div
+      className="App"
+      style={{ backgroundColor: "#f5f5f5", height: "945px", overflowX: "hidden", position: "relative" }}
+    >
       {/* <Navigation /> */}
       <MainNavContainer />
       <MessageBox />
       {isLoading ? <Loading /> : null}
-      <Switch>
-        <Route exact path="/" component={Genres} />
-        <Route path="/signup" component={SignUp} />
-        <Route path="/login" component={Login} />
-        <Route path="/genres/:id" component={GenreDetails} />
-        <Route path="/me" component={MyPage} />
-      </Switch>
+      <AnimatePresence>
+        <Switch location={location} key={location.pathname}>
+          <Route exact path="/" component={Genres} />
+          <Route path="/signup" component={SignUp} />
+          <Route path="/login" component={Login} />
+          <Route path="/genres/:id" component={GenreDetails} />
+          <Route path="/me" component={MyPage} />
+        </Switch>
+      </AnimatePresence>
     </div>
   );
 }
