+++ src/components/Loading/index.js
@@ -2,6 +2,8 @@ import React from "react";
 import "./spinner.css";
 import { makeStyles } from "@material-ui/core/styles";
 import CircularProgress from "@material-ui/core/CircularProgress";
+import { motion } from "framer-motion";
+// import { pageTransitions } from "../../pages/pageTransitions";
 
 const useStyles = makeStyles((theme) => ({
   loader: {
@@ -12,8 +14,8 @@ const useStyles = makeStyles((theme) => ({
 export default function Loading() {
   const classes = useStyles();
   return (
-    <div className="loading_spinner" >
+    <motion.div className="loading_spinner" initial={{ opacity: 0 }} exit={{ opacity: 0 }} animate={{ opacity: 1 }}  >
       <CircularProgress className={classes.loader} />
-    </div>
+    </motion.div>
   );
 }
