+++ src/pages/Genres/DraggingBoard.js
@@ -1,79 +1,9 @@
-import React, { useState, useEffect, useRef } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { fetchGenres } from "../../store/genres/actions";
-import {
-  selectGenres,
-  selectGenreRelations,
-} from "../../store/genres/selectors";
+import React from "react";
 import GenreItem from "./GenreItem";
 import RelationLine from "./RelationLine";
-import Loading from "../../components/Loading";
 
-export default function DraggingBoard() {
-  const [tiles, setTiles] = useState({});
-  const [lines, setLines] = useState([]);
-  const [loading, setLoading] = useState({ loading: "true" });
-  const dispatch = useDispatch();
-  const genres = useSelector(selectGenres);
-  const relations = useSelector(selectGenreRelations);
-  const constraintsRef = useRef(null);
-
-  useEffect(() => {
-    dispatch(fetchGenres());
-  }, [dispatch]);
-
-  // console.log("tiles", tiles);
-  if (!genres) return <h5>Loading</h5>;
-
-  // if (genres.length === 0) return <Loading />;
-  // if (lines.length === 0) return <h5>Loading</h5>;
-
-  if (!tiles[1]) {
-    const top = -20;
-    const left = 60;
-    const addTopLeft = (elements, x, y) =>
-      elements.map((element) => {
-        const newElement = { ...element, top: (x += 40), left: y };
-        return newElement;
-      });
-
-    const genresAddedXY = addTopLeft(genres, top, left);
-
-    const arrayToObject = (array, keyField) =>
-      array.reduce((obj, item) => {
-        obj[item[keyField]] = item;
-        return obj;
-      }, {});
-
-    const initialState = arrayToObject(genresAddedXY, "id");
-
-    setTiles(initialState);
-
-    const addTopLeftLine = (elements) =>
-      elements.map((element) => {
-        const newElement = {
-          ...element,
-          genLX: 0,
-          genTY: 0,
-          othGenLX: 0,
-          othGenTY: 0,
-        };
-        return newElement;
-      });
-
-    let linesArray = addTopLeftLine(relations);
-
-    Object.keys(initialState).forEach((tileId) => {
-      const tile = initialState[tileId];
-
-      linesArray = calculateLines(linesArray, tile);
-      // console.log("lines array", linesArray);
-    });
-    setLines(linesArray);
-    setLoading({ loading: "false" });
-  }
-
-  // console.log("lines", lines);
+export default function DraggingBoard(props) {
+  const { tiles, setTiles, lines, setLines, calculateLines } = props;
 
   const startDragging = (e, id) => {
     const mouseDown = { x: e.clientX, y: e.clientY };
@@ -116,29 +46,6 @@ export default function DraggingBoard() {
     });
   };
 
-  function calculateLines(lines, tile) {
-    return lines.map((line, index) => {
-      // console.log("yeste, isto je genreId", line);
-
-      if (tile.id === line.genreId) {
-        return {
-          ...line,
-          genTY: tile.top + 15,
-          genLX: tile.left + 80,
-        };
-      }
-      if (tile.id === line.otherGenreId) {
-        return {
-          ...line,
-          othGenTY: tile.top + 15,
-          othGenLX: tile.left + 80,
-        };
-      } else {
-        return line;
-      }
-    });
-  }
-
   const stopDragging = () => {
     // console.log("stopped");
     Object.keys(tiles).forEach((tileId) => {
@@ -161,12 +68,7 @@ export default function DraggingBoard() {
   };
 
   return (
-    <div
-      ref={constraintsRef}
-      style={boardStyle}
-      onMouseUp={stopDragging}
-      onMouseMove={onDrag}
-    >
+    <div style={boardStyle} onMouseUp={stopDragging} onMouseMove={onDrag}>
       {Object.keys(tiles).map((id) => {
         const tile = tiles[id];
         return (
@@ -177,27 +79,13 @@ export default function DraggingBoard() {
             key={id}
             setTiles={setTiles}
             tiles={tiles}
-            forwardedRef={constraintsRef}
           />
         );
       })}
 
-      {/* TESTING NEW SVG */}
       <svg height="870" width="100%">
         {lines.map((line) => {
-          return (
-            // <RelationLine />
-            <line
-              strokeWidth="1px"
-              stroke="#3f51b5"
-              x1={line.genLX}
-              y1={line.genTY}
-              x2={line.othGenLX}
-              y2={line.othGenTY}
-              id={line.id}
-              key={line.id}
-            />
-          );
+          return <RelationLine key={line.id} line={line} setLines={setLines} />;
         })}
       </svg>
     </div>
