+++ src/pages/Login/index.js
@@ -3,6 +3,8 @@ import { login } from "../../store/user/actions";
 import { selectToken } from "../../store/user/selectors";
 import { useDispatch, useSelector } from "react-redux";
 import { useHistory, Link } from "react-router-dom";
+import { motion } from "framer-motion"
+import { pageTransitions, pageVariants, pageStyle } from "../pageTransitions";
 
 import {
   Button,
@@ -51,71 +53,73 @@ export default function SignUp() {
   }
 
   return (
-    <React.Fragment>
-      <CssBaseline />
-      <Container maxWidth="xs" style={{ textAlign: "center" }}>
-        <form className={classes.root}>
-          <Box pt={6}>
-            <div
-              style={{
-                backgroundColor: "#a6b3fd",
-                borderRadius: "100%",
-                padding: "7px",
-                width: "40px",
-                margin: "0 auto",
-              }}
-            >
-              <LockOpenIcon />
-            </div>
-          </Box>
-          <Box mb={3}>
-            <h1>Login</h1>
-          </Box>
-          <Box mb={3}>
-            <TextField
-              style={{ width: "100%", margin: "0" }}
-              value={email}
-              onChange={(event) => setEmail(event.target.value)}
-              type="email"
-              placeholder="Enter email"
-              required
-              label="Email address"
-              variant="outlined"
-            />
-          </Box>
-          <Box>
-            <TextField
-              style={{ width: "100%", margin: "0" }}
-              value={password}
-              onChange={(event) => setPassword(event.target.value)}
-              type="password"
-              placeholder="Password"
-              required
-              label="Password"
-              variant="outlined"
-            />
-          </Box>
-          <Box mt={7} mb={1}>
-            <Button
-              className="MuiButtonBase-root MuiButton-root MuiButton-contained jss4 MuiButton-containedPrimary MuiButton-fullWidth"
-              variant="contained"
-              color="primary"
-              type="submit"
-              onClick={submitForm}
-            >
-              Log in
-            </Button>
-          </Box>
-          <Box className="MuiGrid-root MuiGrid-item">
-            <Link
-              className="MuiTypography-root MuiLink-root MuiLink-underlineHover MuiTypography-body2 MuiTypography-colorPrimary"
-              to="/signup"
-            >
-              Don't have an account? Sign Up
-            </Link>
-          </Box>
-        </form>
-      </Container>
-    </React.Fragment>
+    <motion.div style={pageStyle} initial="initial" exit="out" animate="in" variants={pageVariants} transition={pageTransitions}>
+      <React.Fragment>
+        <CssBaseline />
+        <Container maxWidth="xs" style={{ textAlign: "center" }}>
+          <form className={classes.root}>
+            <Box pt={6}>
+              <div
+                style={{
+                  backgroundColor: "#a6b3fd",
+                  borderRadius: "100%",
+                  padding: "7px",
+                  width: "40px",
+                  margin: "0 auto",
+                }}
+              >
+                <LockOpenIcon />
+              </div>
+            </Box>
+            <Box mb={3}>
+              <h1>Login</h1>
+            </Box>
+            <Box mb={3}>
+              <TextField
+                style={{ width: "100%", margin: "0" }}
+                value={email}
+                onChange={(event) => setEmail(event.target.value)}
+                type="email"
+                placeholder="Enter email"
+                required
+                label="Email address"
+                variant="outlined"
+              />
+            </Box>
+            <Box>
+              <TextField
+                style={{ width: "100%", margin: "0" }}
+                value={password}
+                onChange={(event) => setPassword(event.target.value)}
+                type="password"
+                placeholder="Password"
+                required
+                label="Password"
+                variant="outlined"
+              />
+            </Box>
+            <Box mt={7} mb={1}>
+              <Button
+                className="MuiButtonBase-root MuiButton-root MuiButton-contained jss4 MuiButton-containedPrimary MuiButton-fullWidth"
+                variant="contained"
+                color="primary"
+                type="submit"
+                onClick={submitForm}
+              >
+                Log in
+              </Button>
+            </Box>
+            <Box className="MuiGrid-root MuiGrid-item">
+              <Link
+                className="MuiTypography-root MuiLink-root MuiLink-underlineHover MuiTypography-body2 MuiTypography-colorPrimary"
+                to="/signup"
+              >
+                Don't have an account? Sign Up
+              </Link>
+            </Box>
+          </form>
+        </Container>
+      </React.Fragment>
+    </motion.div>
   );
 }
